requirements.txt
streamlit>=1.0
tensorflow>=2.12
pillow
numpy